Fix regarding workaround for a bug in get_area-function
Forgot that I had get_area-function fixed in my own project. This, combined with a fix in the new hydro module halved the area causing drag. This led the drag force to also halve in my tests. Therefore the plate dropped from 15 m using correctly fixed get_area led it to fly to the sky instead of bouncing few times as described.